The server ID is now generated client side using BLAKE2s
This replaces the former functionality where the client would
pass a BLAKE2s HMAC up to the server, which would then SHA256
that value for use as the server ID. This allowed the server
to briefly have knowledge of the client ID, which could
be then used as a selector to find the holder of the secret
link bearing that ID which could be either the sender or the
recipient. If they could find this link, they have the encryption
key as well and all is lost if the DB is compromised.

Now the hashing action is done on the client side and the
server never knows the ID that is in the shared link, and of
course the server still never knows the encrypt/decrypt key
either.

It was considered to use scrypt, instead of a one-way hash,
to generate this server ID, but this idea was discarded due to
the introduction of a new complexity in the need of a salt for
scrypt and the need to then include this salt in the secret link.

This should not decrease the security of the solution since the ID
is not a part of the cryptographic security of the stored secret,
it is merely an identifier and is used for integrity checks. It is
also considered infeasible at this time to reverse the BLAKE2s hash
function by brute forcing hash attempts against known server IDs
which, for example, would be learned in the case of a server breach.